Issue #163 Combined objective and rule panels
Everything is in one class and window now. Serious refactoring to
simplify the code and allow for simpler rule modification. Removed
extra unneeded JPanels that were only being used for adding labels, and
changed all layouts to be gridbag for conistency and simplicity.
Fixed issue #164 that wasn't allowing anything but Kings or Knights as
the objective piece for games.